Refactor: Move FileUtil
