+++ core/src/main/java/bisq/core/app/AvoidStandbyModeService.java
@@ -20,7 +20,7 @@ package bisq.core.app;
 import bisq.core.user.Preferences;
 
 import bisq.common.config.Config;
-import bisq.common.storage.FileUtil;
+import bisq.common.file.FileUtil;
 import bisq.common.storage.ResourceNotFoundException;
 import bisq.common.util.Utilities;
