+++ p2p/src/main/java/bisq/network/p2p/network/TorMode.java
@@ -17,18 +17,18 @@
 
 package bisq.network.p2p.network;
 
-import java.io.File;
-import java.io.IOException;
+import bisq.common.file.FileUtil;
 
 import org.berndpruenster.netlayer.tor.Tor;
 import org.berndpruenster.netlayer.tor.TorCtlException;
 
-import bisq.common.storage.FileUtil;
+import java.io.File;
+import java.io.IOException;
 
 /**
  * Holds information on how tor should be created and delivers a respective
  * {@link Tor} object when asked.
- * 
+ *
  * @author Florian Reimair
  *
  */
@@ -57,7 +57,7 @@ public abstract class TorMode {
 
     /**
      * Returns a fresh {@link Tor} object.
-     * 
+     *
      * @return a fresh instance of {@link Tor}
      * @throws IOException
      * @throws TorCtlException
@@ -72,7 +72,7 @@ public abstract class TorMode {
      * service path literally. Hence, we set <code>"torDir/hiddenservice"</code> as
      * the hidden service directory. By doing so, we use the same
      * <code>private_key</code> file as in {@link NewTor} mode.
-     * 
+     *
      * @return <code>""</code> in {@link NewTor} Mode,
      *         <code>"torDir/externalTorHiddenService"</code> in {@link RunningTor}
      *         mode
